requirements.txt
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap4==4.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
visitor==0.1.3
Werkzeug==0.14.1
